Added Travis CI requirements
First is adding .travis.yml which tells Travis what to do.

Next is adding a Gemfile which specifies the requirements for the projects.
The downside of this is we will now have to use bundle for all commands
relating to the blog (no more jekyll serve). So the first time you work on
the blog you need to do a 'bundle install' to install the dependancies.
From then on you need to prefix your commands with 'bundle exec' so
you can serve with 'bundle exec jekyll serve'.

That however is a bit long winded so we will also have a Rakefile which
contains some commands which can be run by specifying 'rake <command>'.
So instead of 'bundle exec jekyll serve' we can do 'rake serve'. We can
also test the blog by running 'rake test' which is what Travis will be doing.